requirements.txt
pandas==3.0.2
PyDrive2==1.21.3
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
oauth2client==4.1.3
matplotlib==3.10.8
numpy==2.4.4
PyYAML==6.0.3
requests==2.33.1
